Executable builds, check over tests
