+++ tdms/src/arrays.cpp
@@ -10,16 +10,6 @@
 using namespace std;
 using namespace tdms_math_constants;
 
-bool DispersiveMultiLayer::is_dispersive(double near_zero_tolerance) const {
-  for (double gamma_val : gamma) {
-    if (fabs(gamma_val) > near_zero_tolerance) {
-      // non-zero attenuation constant of a Yee cell implies media is dispersive
-      return true;
-    }
-  }
-  return false;
-}
-
 FrequencyExtractVector::FrequencyExtractVector(const mxArray *ptr,
                                                double omega_an) {
